% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{diff_expression}
\alias{diff_expression}
\title{diff_expression}
\usage{
diff_expression(
  counts,
  metadata,
  sample_col = "sample",
  condition_col = "condition",
  condition1,
  condition2
)
}
\arguments{
\item{counts}{Um data frame com os dados de contagem de expressão gênica. As colunas devem ser amostras e as linhas devem ser genes.}

\item{metadata}{Um data frame com informações sobre as amostras, incluindo as condições experimentais. Deve conter pelo menos duas colunas: uma para as amostras e outra para as condições.}

\item{sample_col}{Nome da coluna no data frame \code{metadata} que contém os identificadores das amostras. O padrão é \code{"sample"}.}

\item{condition_col}{Nome da coluna no data frame \code{metadata} que contém as condições experimentais (ex: caso vs controle). O padrão é \code{"condition"}.}

\item{condition1}{Valor da condição de controle para a comparação. Por exemplo, \code{"controle"}.}

\item{condition2}{Valor da condição experimental para a comparação. Por exemplo, \code{"caso"}.}
}
\value{
Uma lista com dois data frames:
\itemize{
\item \code{DEG_Results}: Contém todos os resultados da expressão diferencial, incluindo log fold change, erro padrão, valor p ajustado, e comparação.
\item \code{Comparison_Info}: Contém informações sobre a comparação realizada.
}
}
\description{
Esta função realiza a análise de expressão diferencial usando contagens de RNA-Seq e metadados. Ela calcula as diferenças de expressão entre duas condições e realiza a análise diferencial usando o modelo linear.
}
