% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDEG.R
\name{metaDEG}
\alias{metaDEG}
\title{metaDEG}
\usage{
metaDEG(
  DE_results_list,
  sm = "SMD",
  method.tau = "REML",
  hakn = TRUE,
  method.ci = "z",
  prediction = TRUE,
  fixed = FALSE,
  random = TRUE
)
}
\arguments{
\item{DE_results_list}{Uma lista combinada dos resultados individuais de DEGs gerados pela função diff_expression.}

\item{sm}{Medida de efeito (padrão: "SMD" - Diferença de Média Padronizada).
Outras opções: "MD" (Diferença de Médias), "OR" (Odds Ratio), "RR" (Risk Ratio).}

\item{method.tau}{Método para estimar tau² (padrão: "REML").
Outras opções: "DL" (DerSimonian-Laird), "SJ" (Sidik-Jonkman), "EB" (Empirical Bayes).}

\item{hakn}{Ajuste de variância de Hartung-Knapp (padrão: TRUE).}

\item{method.ci}{Método para calcular intervalos de confiança (padrão: "z").
Outras opções: "t" (intervalos de confiança baseados em t).}

\item{prediction}{Calcular intervalos de predição (padrão: TRUE).}

\item{fixed}{Incluir resultados de efeito fixo (padrão: FALSE).}

\item{random}{Incluir resultados de efeito aleatório (padrão: TRUE).}
}
\value{
Um data frame com os resultados da meta-análise de diferentes experimentos de DEGs.
}
\description{
A função realiza uma meta-análise nos resultados de expressão diferencial combinados de vários estudos.
}
\author{
Bruno Rodrigo Assuncao(2024).
}
